refactor: Trip 대표이미지 url -> name으로 변경
