+++ backend/src/main/java/hanglog/community/dto/response/CommunityTripDetailResponse.java
@@ -4,7 +4,7 @@ import static lombok.AccessLevel.PRIVATE;
 
 import hanglog.city.domain.City;
 import hanglog.city.dto.response.CityWithPositionResponse;
-import hanglog.share.dto.response.WriterResponse;
+import hanglog.member.dto.response.WriterResponse;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.dto.response.DayLogResponse;
 import java.time.LocalDate;
