+++ backend/src/main/java/hanglog/share/dto/response/SharedTripDetailResponse.java
@@ -4,6 +4,7 @@ import static lombok.AccessLevel.PRIVATE;
 
 import hanglog.city.domain.City;
 import hanglog.city.dto.response.CityWithPositionResponse;
+import hanglog.member.dto.response.WriterResponse;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.dto.response.DayLogResponse;
 import java.time.LocalDate;
