+++ backend/src/main/java/hanglog/trip/dto/response/TripDetailResponse.java
@@ -7,7 +7,7 @@ import static lombok.AccessLevel.PRIVATE;
 
 import hanglog.city.domain.City;
 import hanglog.city.dto.response.CityWithPositionResponse;
-import hanglog.share.dto.response.WriterResponse;
+import hanglog.member.dto.response.WriterResponse;
 import hanglog.trip.domain.DayLog;
 import hanglog.trip.domain.Trip;
 import hanglog.trip.dto.response.type.TripType;
@@ -30,7 +30,7 @@ public class TripDetailResponse {
     private final LocalDate startDate;
     private final LocalDate endDate;
     private final String description;
-    private final String imageUrl;
+    private final String imageName;
     private final Boolean isLike;
     private final Long likeCount;
     private final LocalDateTime publishedDate;
